+++ wecube_plugins_itsdangerous/apps/plugin/api.py
@@ -2,118 +2,14 @@
 
 from __future__ import absolute_import
 
-import glob
-import hashlib
 import logging
-import os
-import os.path
-import tempfile
 
-import requests
-from talos.core import config, utils
-from talos.core.i18n import _
+from talos.core import config
 from talos.db import crud
-from talos.utils import scoped_globals
-from wecube_plugins_itsdangerous.common import exceptions, s3
-from wecube_plugins_itsdangerous.common import utils as plugin_utils
 from wecube_plugins_itsdangerous.db import validator as my_validator
-from wecube_plugins_itsdangerous.db import resource
 
 LOG = logging.getLogger(__name__)
 CONF = config.CONF
-
-
-def download_from_url(dir_path, url, random_name=False):
-    filename = url.rsplit('/', 1)[-1]
-    if random_name:
-        filename = '%s_%s' % (utils.generate_uuid(), filename)
-    filepath = os.path.join(dir_path, filename)
-    if url.startswith(CONF.wecube.base_url):
-        # nexus url
-        token = CONF.wecube.token or scoped_globals.GLOBALS.request.auth_token
-        resp = requests.get(url, headers={'Authorization': 'Bearer ' + token}, stream=True)
-        chunk_size = 1024 * 1024
-        stream = resp.raw
-        chunk = stream.read(chunk_size)
-        with open(filepath, 'wb') as f:
-            while chunk:
-                f.write(chunk)
-                chunk = stream.read(chunk_size)
-    else:
-        client = s3.S3Downloader(url)
-        client.download_file(filepath, CONF.wecube.s3.access_key, CONF.wecube.s3.secret_key)
-    return filepath
-
-
-def ensure_url_cached(url):
-    cache_dir = CONF.pakcage_cache_dir
-    filename = url.rsplit('/', 1)[-1]
-    new_filename = hashlib.sha1(url.encode()).hexdigest() + '-' + filename
-    cached_file_path = os.path.join(cache_dir, new_filename)
-    with plugin_utils.lock(new_filename, timeout=300) as locked:
-        if locked:
-            if os.path.exists(cached_file_path):
-                LOG.info('using cache: %s for package: %s', cached_file_path, url)
-            else:
-                with tempfile.TemporaryDirectory() as download_path:
-                    LOG.info('download from: %s for pakcage: %s', url, url)
-                    filepath = download_from_url(download_path, url)
-                    LOG.info('download complete')
-                    os.rename(filepath, cached_file_path)
-        else:
-            raise OSError(_('failed to acquire lock, package cache may not be available'))
-    return cached_file_path, filename
-
-
-class ServiceScript(resource.ServiceScript):
-    def get_contents(self, service, plugin_param):
-        def _file_extension_supported(extensions, filename):
-            for e in extensions:
-                if filename.endswith(e):
-                    return True
-            return False
-
-        # [{"type": None/"sql"/"shell", "content": "...", "name": "additional name info"}]
-        scripts = []
-        script_locations = self.list(filters={'service': service})
-        if len(script_locations) == 0:
-            return scripts
-        script_location = script_locations[0]
-        content_type = script_location['content_type']
-        # if content_field present, use it
-        if script_location['content_field']:
-            content = utils.get_item(plugin_param, script_location['content_field'])
-            if content:
-                scripts.append({'type': content_type, 'content': content, 'name': None})
-        # if endpoint_field present, download from endpoint(s3/artifacts url from platform)
-        if script_location['endpoint_field']:
-            endpoint_url = utils.get_item(plugin_param, script_location['endpoint_field'])
-            filepath, filename = ensure_url_cached(endpoint_url)
-            packed_extensions = [
-                '.zip', '.tar', '.tar.gz', '.tgz', '.tar.bz2', '.tbz2', '.tar.xz', '.txz', '.jar', '.war', '.apk'
-            ]
-            shell_extension = '.sh'
-            sql_extension = '.sql'
-            if _file_extension_supported(packed_extensions, filepath):
-                unzip_path = filepath + '__unpack'
-                plugin_utils.unpack_file(filepath, unzip_path)
-                for name in glob.glob(os.path.join(unzip_path, '**/*' + shell_extension), recursive=True):
-                    if not os.path.isfile(name):
-                        continue
-                    with open(name, 'r') as f:
-                        scripts.append({'type': 'shell', 'content': f.read(), 'name': name[len(unzip_path) + 1:]})
-                for name in glob.glob(os.path.join(unzip_path, '**/*' + sql_extension), recursive=True):
-                    if not os.path.isfile(name):
-                        continue
-                    with open(name, 'r') as f:
-                        scripts.append({'type': 'sql', 'content': f.read(), 'name': name[len(unzip_path) + 1:]})
-            elif filepath.endswith(shell_extension):
-                with open(filepath, 'r') as f:
-                    scripts.append({'type': 'shell', 'content': f.read(), 'name': filename})
-            elif filepath.endswith(sql_extension):
-                with open(filepath, 'r') as f:
-                    scripts.append({'type': 'sql', 'content': f.read(), 'name': filename})
-        return scripts
 
 
 class Box(object):
@@ -172,9 +68,9 @@ class Box(object):
                 'serviceName': service,
                 'servicePath': clean_data.get('servicePath', None),
                 'inputParams': input_param,
-                'scripts': ServiceScript().get_contents(service, input_param),
+                'scripts': processor_api.ServiceScript().get_contents(service, input_param),
                 'entityInstances': entity_instances
             }
             input_results = box.check(detect_data)
-            results.append({'is_danger': True if len(input_results) > 0 else False, 'details': input_results})
+            results.append({'is_danger': len(input_results) > 0, 'details': input_results})
         return results
