+++ wecube_plugins_itsdangerous/apps/processor/api.py
@@ -2,52 +2,200 @@
 
 from __future__ import absolute_import
 
+import glob
 import hashlib
 import json
 import logging
+import os
+import os.path
+import tempfile
 
+import requests
 from talos.common import cache
+from talos.core import config, utils
+from talos.core.i18n import _
 from talos.db import crud
-
+from talos.utils import scoped_globals
 from wecube_plugins_itsdangerous.apps.processor import detector
-from wecube_plugins_itsdangerous.common import reader
-from wecube_plugins_itsdangerous.common import scope
-from wecube_plugins_itsdangerous.common import exceptions
+from wecube_plugins_itsdangerous.common import exceptions, reader, s3, scope
+from wecube_plugins_itsdangerous.common import utils as plugin_utils
 from wecube_plugins_itsdangerous.db import resource
 from wecube_plugins_itsdangerous.db import validator as my_validator
 
 LOG = logging.getLogger(__name__)
+CONF = config.CONF
+
+
+def download_from_url(dir_path, url, random_name=False):
+    '''download file from url(s3/atrifact url)
+
+    :param dir_path: file path to save, eg. /tmp/
+    :type dir_path: str
+    :param url: file url
+    :type url: str
+    :param random_name: generate uuid prefix name, defaults to False
+    :type random_name: bool, optional
+    :return: local file path
+    :rtype: str
+    '''
+    filename = url.rsplit('/', 1)[-1]
+    if random_name:
+        filename = '%s_%s' % (utils.generate_uuid(), filename)
+    filepath = os.path.join(dir_path, filename)
+    if url.startswith(CONF.wecube.base_url):
+        # nexus url
+        token = CONF.wecube.token or scoped_globals.GLOBALS.request.auth_token
+        resp = requests.get(url, headers={'Authorization': 'Bearer ' + token}, stream=True)
+        chunk_size = 1024 * 1024
+        stream = resp.raw
+        chunk = stream.read(chunk_size)
+        with open(filepath, 'wb') as f:
+            while chunk:
+                f.write(chunk)
+                chunk = stream.read(chunk_size)
+    else:
+        client = s3.S3Downloader(url)
+        client.download_file(filepath, CONF.wecube.s3.access_key, CONF.wecube.s3.secret_key)
+    return filepath
+
+
+def ensure_url_cached(url):
+    '''download file form url, use cached file if available
+
+    :param url: file url
+    :type url: str
+    :raises OSError: if timeout waiting for lock
+    :return: local file path, filename
+    :rtype: tuple(str, str)
+    '''
+    cache_dir = CONF.pakcage_cache_dir
+    filename = url.rsplit('/', 1)[-1]
+    new_filename = hashlib.sha1(url.encode()).hexdigest() + '-' + filename
+    cached_file_path = os.path.join(cache_dir, new_filename)
+    with plugin_utils.lock(new_filename, timeout=300) as locked:
+        if locked:
+            if os.path.exists(cached_file_path):
+                LOG.info('using cache: %s for package: %s', cached_file_path, url)
+            else:
+                with tempfile.TemporaryDirectory() as download_path:
+                    LOG.info('download from: %s for pakcage: %s', url, url)
+                    filepath = download_from_url(download_path, url)
+                    LOG.info('download complete')
+                    os.rename(filepath, cached_file_path)
+        else:
+            raise OSError(_('failed to acquire lock, package cache may not be available'))
+    return cached_file_path, filename
 
 
 class Policy(resource.Policy):
+    '''
+    Policy Resource for CRUD
+    '''
     pass
 
 
 class Rule(resource.Rule):
+    '''
+    Rule Resource for CRUD
+    '''
     pass
 
 
 class MatchParam(resource.MatchParam):
+    '''
+    MatchParam Resource for CRUD
+    '''
     pass
 
 
 class Subject(resource.Subject):
+    '''
+    Subject Resource for CRUD
+    '''
     pass
 
 
 class Target(resource.Target):
+    '''
+    Target Resource for CRUD
+    '''
     pass
 
 
 class ServiceScript(resource.ServiceScript):
-    pass
+    '''
+    ServiceScript Resource for CRUD
+    '''
+    def get_contents(self, service, plugin_param):
+        '''get script contents from param according to service definition
+
+        :param service: service name, eg. a/b(c)/d
+        :type service: str
+        :param plugin_param: service param instance
+        :type plugin_param: dict
+        :return: list of {'type': None/'sql'/'shell', 'content': '', 'name': ''}
+        :rtype: list
+        '''
+        def _file_extension_supported(extensions, filename):
+            for e in extensions:
+                if filename.endswith(e):
+                    return True
+            return False
+
+        # [{"type": None/"sql"/"shell", "content": "...", "name": "additional name info"}]
+        scripts = []
+        script_locations = self.list(filters={'service': service})
+        if len(script_locations) == 0:
+            return scripts
+        script_location = script_locations[0]
+        content_type = script_location['content_type']
+        # if content_field present, use it
+        if script_location['content_field']:
+            content = utils.get_item(plugin_param, script_location['content_field'])
+            if content:
+                scripts.append({'type': content_type, 'content': content, 'name': None})
+        # if endpoint_field present, download from endpoint(s3/artifacts url from platform)
+        if script_location['endpoint_field']:
+            endpoint_url = utils.get_item(plugin_param, script_location['endpoint_field'])
+            filepath, filename = ensure_url_cached(endpoint_url)
+            packed_extensions = [
+                '.zip', '.tar', '.tar.gz', '.tgz', '.tar.bz2', '.tbz2', '.tar.xz', '.txz', '.jar', '.war', '.apk'
+            ]
+            shell_extension = '.sh'
+            sql_extension = '.sql'
+            if _file_extension_supported(packed_extensions, filepath):
+                unzip_path = filepath + '__unpack'
+                plugin_utils.unpack_file(filepath, unzip_path)
+                for name in glob.glob(os.path.join(unzip_path, '**/*' + shell_extension), recursive=True):
+                    if not os.path.isfile(name):
+                        continue
+                    with open(name, 'r') as f:
+                        scripts.append({'type': 'shell', 'content': f.read(), 'name': name[len(unzip_path) + 1:]})
+                for name in glob.glob(os.path.join(unzip_path, '**/*' + sql_extension), recursive=True):
+                    if not os.path.isfile(name):
+                        continue
+                    with open(name, 'r') as f:
+                        scripts.append({'type': 'sql', 'content': f.read(), 'name': name[len(unzip_path) + 1:]})
+            elif filepath.endswith(shell_extension):
+                with open(filepath, 'r') as f:
+                    scripts.append({'type': 'shell', 'content': f.read(), 'name': filename})
+            elif filepath.endswith(sql_extension):
+                with open(filepath, 'r') as f:
+                    scripts.append({'type': 'sql', 'content': f.read(), 'name': filename})
+        return scripts
 
 
 class BoxManage(resource.BoxManage):
+    '''
+    Rule Resource for CRUD
+    '''
     pass
 
 
 class Box(resource.Box):
+    '''
+    Rule Resource for CRUD
+    '''
     runall_rules = [
         crud.ColumnValidator(field='operator',
                              rule=my_validator.LengthValidator(0, 63),
@@ -71,7 +219,7 @@ class Box(resource.Box):
                              nullable=False),
     ]
 
-    def _get_rules(self, data, boxes=None):
+    def _get_rules(self, data, boxes=None, without_subject_test=False):
         boxes = boxes or self.list(filters={'policy.enabled': 1, 'subject.enabled': 1})
         rules = {}
         hasher = hashlib.sha256()
@@ -79,71 +227,90 @@ class Box(resource.Box):
         digest = hasher.hexdigest()
         LOG.debug('scope test with data - %s ...', str(data)[:4096])
         for b in boxes:
-            LOG.debug('scope test of box[%s - %s]', b['id'], b['name'])
+            LOG.info('scope test of box[%s - %s]', b['id'], b['name'])
             subject_included = False
-            for target in b['subject']['targets']:
-                target_included = True
-                # target with the same data is cached
-                key = 'scope/target/%s/data/%s' % (target['id'], digest)
-                cached = cache.get(key, 30)
-                if cache.validate(cached):
-                    target_included = cached
-                    LOG.debug('scope test of target[%s - %s]: %s', target['id'], target['name'],
-                              ('accepted' if cached else 'rejected'))
-                else:
-                    LOG.debug('scope test of target[%s - %s]', target['id'], target['name'])
-                    if target['enabled']:
-                        if target['args_scope']:
-                            target_included = scope.JsonScope(target['args_scope']).is_match(data)
-                        else:
-                            target_included = True
-                        if target_included:
-                            LOG.debug('args scope: accepted')
-                            if target['entity_scope']:
-                                target_included = scope.WeCubeScope(target['entity_scope']).is_match(
-                                    data['entityInstances'])
+            if without_subject_test:
+                subject_included = True
+            else:
+                for target in b['subject']['targets']:
+                    target_included = True
+                    # target with the same data is cached
+                    key = 'scope/target/%s/data/%s' % (target['id'], digest)
+                    cached = cache.get(key, 30)
+                    if cache.validate(cached):
+                        target_included = cached
+                        LOG.debug('scope test of target[%s - %s]: %s', target['id'], target['name'],
+                                  ('accepted' if cached else 'rejected'))
+                    else:
+                        LOG.debug('scope test of target[%s - %s]', target['id'], target['name'])
+                        if target['enabled']:
+                            # jsonscope match & exprscope match
+                            if target['args_scope']:
+                                target_included = scope.JsonScope(target['args_scope']).is_match(data)
                             else:
                                 target_included = True
                             if target_included:
-                                LOG.debug('entity scope: accepted')
+                                LOG.debug('args scope: accepted')
+                                if target['entity_scope']:
+                                    target_included = scope.WeCubeScope(target['entity_scope']).is_match(
+                                        data['entityInstances'])
+                                else:
+                                    target_included = True
+                                if target_included:
+                                    LOG.debug('entity scope: accepted')
+                                else:
+                                    LOG.debug('entity scope: rejected')
                             else:
-                                LOG.debug('entity scope: rejected')
+                                LOG.debug('args scope: rejected')
                         else:
-                            LOG.debug('args scope: rejected')
-                    else:
-                        LOG.debug('target: disabled')
-                        target_included = False
-                cache.set(key, target_included)
-                if target_included:
-                    subject_included = True
-                    break
+                            LOG.debug('target: disabled')
+                            target_included = False
+                    cache.set(key, target_included)
+                    # if any target are included, means subject is inlcuded
+                    if target_included:
+                        subject_included = True
+                        break
             if subject_included:
                 # extend box rules(enabled)
+                new_rules_map = {}
                 for rule in b['policy']['rules']:
                     if rule['enabled']:
-                        rules[rule['id']] = rule
-                LOG.debug('scope test of box[%s - %s]: accepted, rules: %s', b['id'], b['name'], list(rules.keys()))
+                        new_rules_map[rule['id']] = rule
+                rules.update(new_rules_map)
+                LOG.info('scope test of box[%s - %s]: accepted, rules: %s', b['id'], b['name'],
+                         list(new_rules_map.keys()))
             else:
-                LOG.debug('scope test of box[%s - %s]: rejected', b['id'], b['name'])
+                LOG.info('scope test of box[%s - %s]: rejected', b['id'], b['name'])
         return list(rules.values())
 
     def _rule_grouping(self, rules):
-        # {'filter': [r1, r2], 'cli': [r3], 'sql/text/fulltext': [rx...]}
+        # {'filter': [r1, r2], 'cli': [r3], 'sql/text/fulltext/..': [rx...]}
         results = {}
         for r in rules:
             rs = results.setdefault(r['match_type'], [])
             rs.append(r)
         return results
 
-    def run(self, data, rid):
-        refs = Box().list({'id': rid})
+    def simplerun(self, data, rid):
+        '''run box[rid] rule check
+
+        :param data: see function check
+        :type data: dict
+        :param rid: box id
+        :type rid: any
+        :raises exceptions.NotFoundError: if id of box not found
+        :return: see function check
+        :rtype: see function check
+        '''
+        refs = self.list({'id': rid})
         if len(refs) == 0:
             raise exceptions.NotFoundError(resource='Box')
-        return self.check(data, boxes=refs)
+        return self.check(data, boxes=refs, without_subject_test=True)
 
-    def runall(self, data):
-        '''
-        input data:
+    def plugin_check(self, data):
+        '''run plugin params check
+
+        :param data: input data
         {
             // 操作人
             "operator": "admin",
@@ -157,8 +324,10 @@ class Box(resource.Box):
                 {}
             ]
         }
+
+        :return: see funcion check
+        :rtype: see funcion check
         '''
-        from wecube_plugins_itsdangerous.apps.plugin import api as plugin_api
         results = []
         clean_data = crud.ColumnValidator.get_clean_data(self.runall_rules, data, 'check')
         service = clean_data['serviceName']
@@ -169,26 +338,33 @@ class Box(resource.Box):
                 'serviceName': service,
                 'servicePath': clean_data.get('servicePath', None),
                 'inputParams': input_param,
-                'scripts': plugin_api.ServiceScript().get_contents(service, input_param),
+                'scripts': ServiceScript().get_contents(service, input_param),
                 'entityInstances': entity_instances
             }
             input_results = self.check(detect_data)
-            results.append({'is_danger': True if len(input_results) > 0 else False, 'details': input_results})
+            results.extend(input_results)
         return results
 
-    def check(self, data, boxes=None):
-        '''
-        data: {
-            (Optional - JsonScope check)"serviceName": "a/b(c)/d", 
-            (Optional - JsonScope check)"servicePath": "a/b/run", 
-            (Optional - JsonScope check)"inputParams": {...service input params}, 
-            (Must - script check)"scripts": [{"type": None/"sql"/"shell", "content": "...", "name": "additional name info"}], 
-            (Must - WeCMDBScope check)"entityInstances": [{"guid": "xxx_xxxxxx"}]}
+    def check(self, data, boxes=None, without_subject_test=False):
+        '''check script & param with boxes, return dangerous contents & rule name
+
+        :param data: data with param & script content
+        {
+            (Optional - JsonScope check)"serviceName": "a/b(c)/d",
+            (Optional - JsonScope check)"servicePath": "a/b/run",
+            (Optional - JsonScope check)"inputParams": {...service input params},
+            (Must - script check)"scripts": [{"type": None/"sql"/"shell", "content": "...", "name": "filename"}],
+            (Must - WeCubeScope check)"entityInstances": [{"id": "xxx_xxxxxx"}]}
+        :type data: dict
+        :param boxes: specific boxes if any, defaults to None, mean all boxes
+        :type boxes: list of Box, optional
+        :return: list of {'lineno': [start, end], 'level': level of rule, 'content': content, 'message': rule name}
+        :rtype: list
         '''
 
         results = []
         scripts = data['scripts']
-        rules = self._get_rules(data, boxes=boxes)
+        rules = self._get_rules(data, boxes=boxes, without_subject_test=without_subject_test)
         rules = self._rule_grouping(rules)
         for item in scripts:
             script_name = item.get('name', '') or ''
