+++ wecube_plugins_itsdangerous/apps/processor/route.py
@@ -28,4 +28,4 @@ def add_routes(api):
     api.add_route('/itsdangerous/ui/v1/boxes', controller.CollectionBox())
     api.add_route('/itsdangerous/ui/v1/boxes/{rid}', controller.ItemBox())
     api.add_route('/itsdangerous/ui/v1/boxes/{rid}/run', controller.BoxRun())
-    api.add_route('/itsdangerous/v1/batch_execution_detection', controller.Run())
+    api.add_route('/itsdangerous/v1/batch_execution_detection', controller.PluginCheck())
